Reaction to issues #2 - #4
As a start the code is now in PEP8 format.
The code was changed in the Metrastat.py file so that it can be imported
in the Metrastat.ipynb module rather than to duplicate the code again.
The two .json files contain the constants and initial values used in the
differential model. The parameters.py module contains functions that
will read these values from the .json files and return them as a single
array.