First fitting of experimental data
I made the following changes to existing files:
*Metrastat.py: The code to plot the data at the end of the script was
unnecessary and unwanted when the metrastat function is called in other
scripts. I removed it from the module. Furthermore the HCl removal rate
constant was initially a non-compulsory input. I changed it to be a
required input seeing that it will now be determined using a
least-squares fit.
*Parameters.py: Only k3-k8 is actually used in the model. The unused
constants were removed.

I added the following files:
*data_refine.py: This module will extract the file directories of all
the required .csv files so that in can be used as an input to the
panda.read_csv function. It also contains a function to remove the
inconsistencies in the experimental data so that relevant data may be
fitted using a least-squares algorithm.
*fitting_YI.py: The lmfit module is used to determine the value of the
HCl removal rate constant such that it will minimize the error between
the model fit and the actual experimental data.
*least_squares_fit.py: A value for the HCl removal rate constant was
assumed after which a least-squares fitting was done on a sample of
experimental data